feat(radio-specs): add poetry run update-radio-specs command
Move the extraction logic into veaf_build.radio_specs_updater and register
it as a Poetry script entry point. The scripts/ wrapper becomes a thin
delegator. Update all doc references to use the new command.
