+++ doc/PIPELINE_REFERENCE.md
@@ -149,7 +149,7 @@ La vérification est **non bloquante** : la mission est quand même écrite, mai
 Les specs couvrent 85 aéronefs pilotables et sont issues de [dcs-lua-datamine](https://github.com/Quaggles/dcs-lua-datamine). Si un aéronef n'est pas dans la base, la vérification est silencieusement ignorée.
 
 > **Voir aussi** : [`doc/mission-maker/dcs-radio-specs.md`](dcs-radio-specs.md) — table de référence complète des plages de fréquences valides par aéronef.  
-> Pour régénérer après une mise à jour DCS : `python scripts/extract_dcs_radio_specs.py`
+> Pour régénérer après une mise à jour DCS : `poetry run update-radio-specs`
 
 ---
